+++ src/arena/components/render.rs
